// src/services/cep/types.ts
export interface ICepPayload {
  cep?: string;
  logradouro?: string;
  complemento?: string;
  unidade?: string;
  bairro?: string;
  localidade?: string;
  uf?: string;
  estado?: string;
  regiao?: string;
  ibge?: string;
  gia?: string;
  ddd?: string;
  siafi?: string;
  code?: string;
  state?: string;
  city?: string;
  district?: string;
  address?: string;
  address_type?: string;
  address_name?: string;
  lat?: string;
  lng?: string;
  city_ibge?: string;
}